+++ apps/coordinator/src/actions/keystoreActions.ts
@@ -1,4 +1,5 @@
 import {
+  JADE,
   BITBOX,
   TREZOR,
   LEDGER,
@@ -12,6 +13,7 @@ export const SET_KEYSTORE_NOTE = "SET_KEYSTORE_NOTE";
 export const SET_KEYSTORE_STATUS = "SET_KEYSTORE_STATUS";
 
 type KeyStoreType =
+  | typeof JADE
   | typeof BITBOX
   | typeof TREZOR
   | typeof LEDGER