+++ apps/coordinator/src/components/ScriptExplorer/SignatureImporter.jsx
@@ -9,7 +9,14 @@ import {
   getMaskedDerivation,
   P2SH,
 } from "@caravan/bitcoin";
-import { BITBOX, TREZOR, LEDGER, HERMIT, COLDCARD } from "@caravan/wallets";
+import {
+  JADE,
+  BITBOX,
+  TREZOR,
+  LEDGER,
+  HERMIT,
+  COLDCARD,
+} from "@caravan/wallets";
 import {
   Card,
   CardHeader,
@@ -123,6 +130,7 @@ class SignatureImporter extends React.Component {
             <MenuItem value={COLDCARD} disabled={!isWallet}>
               Coldcard
             </MenuItem>
+            <MenuItem value={JADE}>Jade</MenuItem>
             <MenuItem value={HERMIT}>Hermit</MenuItem>
             <MenuItem value={BCUR2}>BCUR2</MenuItem>
             <MenuItem value={TEXT}>Enter as text</MenuItem>
@@ -155,7 +163,12 @@ class SignatureImporter extends React.Component {
     } = this.props;
     const { method } = signatureImporter;
 
-    if (method === BITBOX || method === TREZOR || method === LEDGER) {
+    if (
+      method === BITBOX ||
+      method === TREZOR ||
+      method === LEDGER ||
+      method === JADE
+    ) {
       return (
         <DirectSignatureImporter
           network={network}