+++ apps/coordinator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.13.0
+
+### Minor Changes
+
+- [#324](https://github.com/caravan-bitcoin/caravan/pull/324) [`531186a`](https://github.com/caravan-bitcoin/caravan/commit/531186a7031961103752f612566a497557a5678f) Thanks [@Austin-Fulbright](https://github.com/Austin-Fulbright)! - Add support for the jade hardware wallet
+
+### Patch Changes
+
+- Updated dependencies [[`531186a`](https://github.com/caravan-bitcoin/caravan/commit/531186a7031961103752f612566a497557a5678f)]:
+  - @caravan/wallets@0.9.0
+
 ## 1.12.1
 
 ### Patch Changes